+++ packages/toast/src/toast.vue
@@ -20,8 +20,8 @@
 </template>
 
 <script>
-import vanLoading from 'packages/loading';
-import vanIcon from 'packages/icon';
+import vanLoading from '../../loading';
+import vanIcon from '../../icon';
 
 const TOAST_TYPES = ['text', 'html', 'loading', 'success', 'fail'];
 const DEFAULT_STYLE_LIST = ['success', 'fail'];